Moved login and status functions to modules.
This should help with code re-use and maintenance.
As we add more functionality, we can keep the individual files to a
minimum, manageable size and if any depend on each other, import as
necessary with statements like `from module import function`.
FYI, if get this error when using separate functions:
TypeError: 'module' object is not callable
Check out these links to understand the why.
https://careerkarma.com/blog/python-typeerror-module-object-is-not-callable/
https://stackoverflow.com/questions/4534438/typeerror-module-object-is-not-callable