Basic logic to pick a filer to enable telnet on.
Very basic for now but when picking the unlock task, the module
will print all connected filers, and relevant details including
the tenant name. After printing, the user is prompted to enter a
device name and then the tenant name to pass to the enable() for
turning on telnet. If the device name or tenant is not found, a
traceback is printed which we should add logic to handle soon.