Added info func, removed print from unlock, added error handling.
Removed the for loop that printed the devie details.
It's not needed to unlock and can be made into its own task if needed.
Created a simple info() function to print the needed info for support.
Raise an exception if the device name entered is not found.k

TODO: Re-add the logic that checks that the device is connected (maybe).