Exit if there's an exception during login.
This will cause ctools to exit instead of continuing on to the
selected task even if the login failed. Works well if the portal
input is left blank or incorrect username or password is used.
The default timeout for an incorrect portal URL though still seems
long so that might be something to add to the #todo list.