Change from menu driven approach to arguments.
Relates to #17 but mostly breaks the old menu function.
I wanted to add the option to pass arguments which would work towards a
goal of being able to script some of these tasks.
Specifically went with using argparse because it's a standard Python
library and may let us use Gooey for #20, to more easily add a GUI.