Fixed ctools menu and task args still work.
As part of work for #17, I broke the menu after adding the option to
pass tasks as arguments. That is fixed now so running ctools with no
args or with `--menu`` flag will show the menu but args still work to
call each task as well.

Decided to remove returning to the menu after finishing `get_status`
or `enable_telnet` functions. With arguments being added, that adds
scripting potential to ctools so the menu has to be an optional thing.