Get Filer status per tenant or globally.
Updated docstring in ctools.py, added optional --all flag which
will run the previoiusly default function of getting all connected
filers to a portal globally which is slow for large deployments.
Commented out the function to browse the admin portal.
By doing the above, we will now default to browsing the tenant
based on the current session and only browse to Administration if
the --all flag is set which sets all_tenants to True in filer.py.

In filer.py, I added tenant name column and checking if the output
file already exists, if it does, we skip writing the header and
just append to the file. I also added try statements to some values
that were throwing exceptions on a 6.0.247 filer.